+++ src/model/toJSONLD.js
@@ -7,6 +7,10 @@
  */
 export default function toJSONLD( variable ) {
 
+  // keep track of blank node identifiers
+  /** @type {Map<Concept, string>} */
+  const blankNodeMap = new Map();
+
   // basic structure
   const result = {
     '@context': 'https://w3id.org/iadopt/Variable.context.jsonld',
@@ -17,15 +21,13 @@ export default function toJSONLD( variable ) {
     comment:  variable.getComment(),
 
     property: {
+      ... serializeEntity( variable.getProperty(), blankNodeMap ),
       '@type': [ 'https://w3id.org/iadopt/ont/Property' ],
-      '@id':    variable.getProperty()?.getIri(),
-      label:    variable.getProperty()?.getLabel(),
-      comment:  variable.getProperty()?.getComment(),
     },
 
-    statisticalModifier:  serializeEntity( variable.getStatisticalModifier() ),
-    ooi:                  serializeEntity( variable.getObjectOfInterest() ),
-    matrix:               serializeEntity( variable.getMatrix() ),
+    statisticalModifier:  serializeEntity( variable.getStatisticalModifier(), blankNodeMap ),
+    ooi:                  serializeEntity( variable.getObjectOfInterest(), blankNodeMap ),
+    matrix:               serializeEntity( variable.getMatrix(), blankNodeMap ),
 
     context: variable.getContextObjects()?.map( (ctx) => serializeEntity( ctx ) ),
     constraint: [],
@@ -33,23 +35,13 @@ export default function toJSONLD( variable ) {
   };
 
   for( const c of variable.getConstraints() ) {
-    if( c.isBlank() ) {
-      result.constraint.push({
-        '@type': [ 'https://w3id.org/iadopt/ont/Constraint' ],
-        label:      c.getLabel(),
-        comment:    c.getComment(),
-        constrains: c.getEntities()?.map( (e) => e.getIri() ),
-      });
-
-    } else {
-      result.constraint.push({
-        '@type': [ 'https://w3id.org/iadopt/ont/Constraint' ],
-        '@id':      c.getIri(),
-        label:      c.getLabel(),
-        comment:    c.getComment(),
-        constrains: c.getEntities()?.map( (e) => e.getIri() ),
-      });
-    }
+    result.constraint.push({
+      '@type': [ 'https://w3id.org/iadopt/ont/Constraint' ],
+      '@id':      c.isBlank() ? undefined : c.getIri(),
+      label:      c.getLabel(),
+      comment:    c.getComment(),
+      constrains: c.getEntities()?.map( (e) => blankNodeMap.has( e ) ? blankNodeMap.get( e ) : e.getIri() ),
+    });
   };
 
   return result;
@@ -58,22 +50,32 @@ export default function toJSONLD( variable ) {
 
 /**
  * Serialize a single Entity.
- * Entity may be a System. Here, all subcomponents are serialized as well
+ * Entity may be a System. Here, all sub-components are serialized as well
  *
  * @param   {import('./models').Entity}   ent
+ * @param   {Map<Concept, string>}        blankNodeMap
  * @returns {Object|void}
  */
-function serializeEntity( ent ) {
+function serializeEntity( ent, blankNodeMap ) {
 
   // just return for empty object
   if( !ent ) {
     return;
   }
 
+  // determine IRI
+  let iri = ent.getIri();
+  if( ent.isBlank() ) {
+    if( !blankNodeMap.has( ent ) ) {
+      blankNodeMap.set( ent, `_:b${blankNodeMap.size}` );
+    }
+    iri = blankNodeMap.get( ent );
+  }
+
   // basic structure
   const result = {
     '@type': [ 'https://w3id.org/iadopt/ont/Entity' ],
-    '@id':    ent.getIri(),
+    '@id':    iri,
     label:    ent.getLabel(),
     comment:  ent.getComment(),
   };
@@ -93,8 +95,8 @@ function serializeEntity( ent ) {
     // add components
     for( const [key, values] of Object.entries( components ) ) {
       result[ key ] = values.length > 1
-        ? values.map( (v) => serializeEntity( v ) )
-        : serializeEntity( values[0] );
+        ? values.map( (v) => serializeEntity( v, blankNodeMap ) )
+        : serializeEntity( values[0], blankNodeMap );
     }
   }