+++ Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## unreleased
+
+### Bugfixes
+* (De)serializing of blank node Entities with Constrains did not keep the proper link
+
 ## 0.3.0
 
 * first public release